Update: cant coerce PubChem to int because stupid Numpy doesn't have NA integer... using float instead

+++ PharmacoDI/build_primary_pset_tables.py
@@ -127,7 +127,7 @@ def build_gene_annotation_df(pset_dict):
 def build_compound_annotation_df(
     pset_dict: dict, 
     column_dict: dict={'compound_id': str, 'smiles': str, 'inchikey': str, 
-        'pubchem': int, 'FDA': bool},
+        'pubchem': float, 'FDA': bool},
     rename_dict: dict={'rownames': 'compound_id', 'cid': 'pubchem', 
             'FDA': 'fda_status'}
 ) -> pd.DataFrame:
